Fix renaming in imports part 3
