+++ src/lexedata/importer/excel_matrix.py
@@ -28,7 +28,7 @@ from lexedata.util.excel import (
     get_cell_comment,
 )
 import lexedata.importer.cellparser as cell_parsers
-from lexedata.enrich.add_status_column import add_status_column_to_table
+from lexedata.edit.add_status_column import add_status_column_to_table
 import lexedata.cli as cli
 
 Ob = t.TypeVar("O", bound=Object)
